api/requirements.txt
sqlalchemy
psycopg2-binary
flask
requests
loguru
BeautifulSoup4